+++ sphinx_design_elements/extension.py
@@ -15,12 +15,15 @@ from .tag import setup_tags
 
 
 def setup_extension(app: Sphinx) -> None:
-    """Set up the sphinx extension."""
+    """
+    Set up the sphinx extension.
+    """
 
-    app.connect("builder-inited", update_css_js)
-    app.connect("env-updated", update_css_links)
-    # we override container html visitors, to stop the default behaviour
-    # of adding the `container` class to all nodes.container
+    app.connect("builder-inited", add_assets)
+    app.connect("env-updated", update_asset_links)
+
+    # Override container html visitors, to stop the default behaviour
+    # of adding the `container` class to all `nodes.container`.
     app.add_node(nodes.container, override=True, html=(visit_container, depart_container))
 
     setup_gridtable(app)
@@ -29,32 +32,52 @@ def setup_extension(app: Sphinx) -> None:
     setup_dropdown_group(app)
 
 
-def update_css_js(app: Sphinx):
-    """Copy the CSS to the build directory."""
-    # reset changed identifier
-    app.env.settings["sphinx_design_elements_css_changed"] = False
-    # setup up new static path in output dir
+def add_assets(app: Sphinx):
+    """
+    Copy JS and CSS assets to the build directory.
+    """
+
+    # Reset "changed" identifier.
+    app.env.settings["sphinx_design_elements_assets_changed"] = False
+
+    # Set up new static path in output dir.
     static_path = (Path(app.outdir) / "_sphinx_design_elements_static").absolute()
     static_existed = static_path.exists()
     static_path.mkdir(exist_ok=True)
     app.config.html_static_path.append(str(static_path))
-    # Read the css content and hash it
-    content = read_text(static_module, "style.css")
+
+    # Add asset files.
+    add_asset(app, static_path, "style.css")
+
+    if not static_existed:
+        app.env.settings["sphinx_design_elements_assets_changed"] = True
+
+
+def add_asset(app: Sphinx, static_path: Path, filename: str):
+    """
+    Copy asset file to the build directory.
+    """
+    content = read_text(static_module, filename)
+    extension = Path(filename).suffix.lstrip(".")
     digest = hashlib.md5(content.encode("utf8")).hexdigest()  # noqa: S324
-    # Write the css file
-    css_path = static_path / f"elements-style.{digest}.min.css"
-    app.add_css_file(css_path.name)
-    if css_path.exists():
-        return
-    if static_existed:
-        app.env.settings["sphinx_design_elements_css_changed"] = True
-    for path in static_path.glob("*.css"):
-        path.unlink()
-    css_path.write_text(content, encoding="utf8")
+    path = static_path / f"design-elements.{digest}.{extension}"
+    if extension == "js":
+        app.add_js_file(path.name)
+    elif extension == "css":
+        app.add_css_file(path.name)
+    else:
+        raise ValueError(f"Unable to add asset file, unknown extension: {filename}")
+    if not path.exists():
+        app.env.settings["sphinx_design_elements_assets_changed"] = True
+        for path_delete in static_path.glob(f"*.{extension}"):
+            path_delete.unlink()
+        path.write_text(content, encoding="utf8")
 
 
-def update_css_links(app: Sphinx, env: BuildEnvironment):
-    """If CSS has changed, all files must be re-written, to include the correct stylesheets."""
-    if app.env.settings.get("sphinx_design_elements_css_changed"):
+def update_asset_links(app: Sphinx, env: BuildEnvironment):
+    """
+    If JS or CSS assets changed, rewrite all files.
+    """
+    if app.env.settings.get("sphinx_design_elements_assets_changed"):
         return list(env.all_docs.keys())
     return []